_apicheck/requirements.txt
PyYAML
user_agent
aiohttp==3.5.4
Cython==0.29.6
Faker==1.0.4
Jinja2==2.10
mitmproxy==4.0.4
numpy==1.16.2
pandas==0.24.2
pg8000==1.13.1
pytest==4.3.1
python-slugify==3.0.1
requests==2.21.0
SQLAlchemy==1.3.1
sqlalchemy-aio==0.14.0
sqlalchemy-stubs==0.1
terminaltables==3.1.0
